api support, refactoring cache code

+++ app/controllers/user_sessions_controller.rb
@@ -50,24 +50,29 @@ class UserSessionsController < ApplicationController
   end
 
   def mobile_create
-      @user_session = UserSession.new(:email => params[:email], :password => params[:password], :remember_me => false)
-      if @user_session.save
-        render :json => { :user_credentials => @user_session.record.single_access_token,
-                          :user_id =>  @user_session.record.id,
-                          :username => @user_session.record.username 
-        }
+    begin
+      user_session = UserSession.new(:email => params[:email], :password => params[:password], :remember_me => false)
+      if user_session.save
+        render :json => JSON.fast_generate({ :user_credentials => user_session.record.persistence_token,
+                          :user_id =>  user_session.record.id,
+                          :username => user_session.record.username
+                        })
       else
-        render_json_error( nil, @user_session.errors.full_messages, 401 )
+        render_json_error( nil, user_session.errors.full_messages, 401 )
       end
+    rescue Exception => ex
+      render_json_error( ex ) and return
+    end
   end
 
   def mobile_destroy
-      if current_user
-          current_user_session.destroy
-      end
+    begin
+      current_user_session.destroy if current_user
       head :status => 200
+    rescue Exception => ex
+      render_json_error( ex ) and return
+    end
   end
-
